# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Labelling Tool", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/projects/": {
      "get": {
        "tags": ["projects"],
        "summary": "Get Own Projects",
        "description": "Get projects owned by the current user.",
        "operationId": "projects-get_own_projects",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectPublic"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Create Project",
        "description": "Create a new project.",
        "operationId": "projects-create_project",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "tags": ["projects"],
        "summary": "Delete Project",
        "description": "Delete a project by ID.",
        "operationId": "projects-delete_project",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/dashboard": {
      "get": {
        "tags": ["projects"],
        "summary": "Get Dashboard Data",
        "description": "Get dashboard statistics for all projects.",
        "operationId": "projects-get_dashboard",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectDashboard"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/dashboard_user": {
      "get": {
        "tags": ["projects"],
        "summary": "Get User Dashboard Data",
        "description": "Get dashboard statistics for current user projects.",
        "operationId": "projects-get_dashboard_user",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectUserDashboard"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/download": {
      "post": {
        "tags": ["projects"],
        "summary": "Download Project",
        "description": "Download project data as JSONL file.",
        "operationId": "projects-download_project",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/status": {
      "get": {
        "tags": ["projects"],
        "summary": "Get Project Status",
        "description": "Get the status of a specific project.",
        "operationId": "projects-get_project_status",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStatus"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/samples": {
      "get": {
        "tags": ["projects"],
        "summary": "Get Line Items",
        "description": "Get line items from a project with pagination and filtering.",
        "operationId": "projects-get_line_items",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["UNLABELED", "CONFIRMED", "APPROVED", "REJECTED"],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/samples/{sample_idx}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get Sample by Index",
        "description": "Get a specific sample by index from a project.",
        "operationId": "projects-get_sample_by_index",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "sample_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sample Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemRead"
                }
              }
            }
          },
          "404": {
            "description": "Sample not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/assign": {
      "post": {
        "tags": ["projects"],
        "summary": "Assign Task",
        "description": "Assign task to a user for a project.",
        "operationId": "projects-assign_task",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/confirm/{line_item_id}": {
      "post": {
        "tags": ["projects"],
        "summary": "Confirm Line Item",
        "description": "Confirm a line item with user feedback.",
        "operationId": "projects-confirm_line_item",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "line_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Line Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/update/{line_item_message_id}": {
      "post": {
        "tags": ["projects"],
        "summary": "Update Line Item Message",
        "description": "Update a line item message with new content and role.",
        "operationId": "projects-update_line_item_message",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "line_item_message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Line Item Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemMessageUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemMessageRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
...
      "ProjectPublic": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 255 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "url": {
            "type": "string",
            "maxLength": 255,
            "title": "Url"
          },
          "owner_id": { "type": "integer", "title": "Owner Id" }
        },
        "type": "object",
        "required": ["id", "name", "url", "owner_id"],
        "title": "ProjectPublic"
      },
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 255 },
              { "type": "null" }
            ],
            "title": "Description"
          },
          "url": {
            "type": "string",
            "maxLength": 255,
            "title": "Url"
          }
        },
        "type": "object",
        "required": ["name", "url"],
        "title": "ProjectCreate"
      },
      "ProjectStatus": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State"
          },
          "info": {
            "anyOf": [{ "type": "object" }, { "type": "null" }],
            "title": "Info"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "num_samples": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Num Samples"
          },
          "num_task_assigned": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Num Task Assigned"
          },
          "num_task_not_assigned": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Num Task Not Assigned"
          },
          "user_task_summary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTaskSummary"
            },
            "title": "User Task Summary"
          }
        },
        "type": "object",
        "required": ["state"],
        "title": "ProjectStatus"
      },
      "AssignTaskRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "num_samples": {
            "type": "integer",
            "minimum": 1,
            "title": "Num Samples"
          }
        },
        "type": "object",
        "required": ["user_id", "num_samples"],
        "title": "AssignTaskRequest"
      },
...
      "LineItemConfirmRequest": {
        "properties": {
          "line_messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineItemMessageConfirmRequest"
            },
            "title": "Line Messages"
          },
          "tools": {
            "anyOf": [
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              { "type": "null" }
            ],
            "title": "Tools"
          },
          "feedback": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Feedback"
          },
          "status": {
            "type": "string",
            "enum": ["UNLABELED", "CONFIRMED", "APPROVED", "REJECTED"],
            "title": "Status"
          }
        },
        "type": "object",
        "required": ["line_messages", "status"],
        "title": "LineItemConfirmRequest"
      },
      "LineItemMessageUpdateRequest": {
        "properties": {
          "role": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Role"
          },
          "content": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Content"
          }
        },
        "type": "object",
        "title": "LineItemMessageUpdateRequest"
      },
      "LineItemMessageRead": {
        "properties": {
          "line_message_index": {
            "type": "integer",
            "title": "Line Message Index"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "feedback": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Feedback"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": ["line_message_index", "role", "content", "id"],
        "title": "LineItemMessageRead"
      },
      "LineItemRead": {
        "properties": {
          "tools": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "title": "Tools"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "line_index": {
            "type": "integer",
            "title": "Line Index"
          },
          "feedback": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Feedback"
          },
          "line_messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineItemMessageRead"
            },
            "title": "Line Messages"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["UNLABELED", "CONFIRMED", "APPROVED", "REJECTED"]
              },
              { "type": "null" }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": ["tools", "id", "project_id", "line_index"],
        "title": "LineItemRead"
      },
      "LineItemsPublic": {
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineItemRead"
            },
            "title": "Data"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "num_pages": {
            "type": "integer",
            "title": "Num Pages"
          },
          "status_counts": {
            "type": "object",
            "properties": {
              "UNLABELED": {
                "type": "integer",
                "title": "Unlabeled Count"
              },
              "CONFIRMED": {
                "type": "integer",
                "title": "Confirmed Count"
              },
              "APPROVED": {
                "type": "integer",
                "title": "Approved Count"
              },
              "REJECTED": {
                "type": "integer",
                "title": "Rejected Count"
              }
            },
            "title": "Status Counts"
          }
        },
        "type": "object",
        "required": ["data", "total_count", "num_pages"],
        "title": "LineItemsPublic"
      },
      "UserTaskSummary": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "task_count": {
            "type": "integer",
            "title": "Task Count"
          },
          "confirmed": {
            "type": "integer",
            "title": "Confirmed"
          },
          "unlabeled": {
            "type": "integer",
            "title": "Unlabeled"
          },
          "approved": {
            "type": "integer",
            "title": "Approved"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "full_name",
          "email",
          "task_count",
          "confirmed",
          "unlabeled",
          "approved",
          "rejected"
        ],
        "title": "UserTaskSummary"
      },
      "ProjectDashboard": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "project_description": {
            "type": "string",
            "title": "Project Description"
          },
          "num_samples": {
            "type": "integer",
            "title": "Num Samples"
          },
          "user_task_summary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTaskSummary"
            },
            "title": "User Task Summary"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "project_name",
          "project_description",
          "num_samples",
          "user_task_summary"
        ],
        "title": "ProjectDashboard"
      },
      "ProjectUserDashboard": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "project_description": {
            "type": "string",
            "title": "Project Description"
          },
          "task_count": {
            "type": "integer",
            "title": "Task Count"
          },
          "status_counts": {
            "type": "object",
            "properties": {
              "UNLABELED": {
                "type": "integer",
                "title": "Unlabeled Count"
              },
              "CONFIRMED": {
                "type": "integer",
                "title": "Confirmed Count"
              },
              "APPROVED": {
                "type": "integer",
                "title": "Approved Count"
              },
              "REJECTED": {
                "type": "integer",
                "title": "Rejected Count"
              }
            },
            "title": "Status Counts"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "project_name",
          "project_description",
          "task_count",
          "status_counts"
        ],
        "title": "ProjectUserDashboard"
      },
      "ProjectDownloadRequest": {
        "properties": {
          "limit": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Limit"
          },
          "include_statuses": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["UNLABELED", "CONFIRMED", "APPROVED", "REJECTED"]
            },
            "title": "Include Statuses"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          }
        },
        "type": "object",
        "required": ["file_name"],
        "title": "ProjectDownloadRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [{ "type": "string" }, { "type": "integer" }]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}